Extract to selectionController.ts
